requirements.txt
vdf = "*"
requests = "*"
beautifulsoup4 = "*"
pysimplegui = "*"
pillow = "*"
pyinstaller = "*"
psgcompiler = "*"
